+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## v0.0.4
+
+### Changed
+* Use pydantic-settings to define global configs
+
+### Added
+* [Auth] JWT auth
+* [Errors] Centralized error handling
+* [CompanionAuthService] External auth service, used to create JWT
+* [Health] Simple health endpoint
+
+
 ## v0.0.3
 
 ### Added